+++ sdk/src/client/secret/ledger_nano.rs
@@ -31,7 +31,7 @@ use crate::{
         output::Output,
         payload::{transaction::TransactionEssence, Payload},
         signature::{Ed25519Signature, Signature},
-        unlock::{AliasUnlock, NftUnlock, ReferenceUnlock, Unlock, Unlocks},
+        unlock::{AliasUnlock, NftUnlock, ReferenceUnlock, SignatureUnlock, Unlock, Unlocks},
     },
     utils::unix_timestamp_now,
 };
@@ -172,8 +172,56 @@ impl SecretManage for LedgerSecretManager {
         Err(Error::UnsupportedOperation.into())
     }
 
-    async fn sign_ed25519(&self, _msg: &[u8], _chain: &Chain) -> Result<Ed25519Signature, Self::Error> {
-        Err(Error::UnsupportedOperation.into())
+    /// Ledger only allows signing messages of 32 bytes, anything else is unsupported and will result in an error.
+    async fn sign_ed25519(&self, msg: &[u8], chain: &Chain) -> Result<Ed25519Signature, Self::Error> {
+        if msg.len() != 32 {
+            return Err(Error::UnsupportedOperation.into());
+        }
+
+        let msg = msg.to_vec();
+
+        let bip32_chain = chain
+            .segments()
+            .iter()
+            // XXX: "ser32(i)". RTFSC: [crypto::keys::slip10::Segment::from_u32()]
+            .map(|seg| u32::from_be_bytes(seg.bs()))
+            .collect::<Vec<_>>();
+
+        let coin_type = bip32_chain[1] & !Segment::HARDEN_MASK;
+        let account_index = bip32_chain[2] | Segment::HARDEN_MASK;
+        let bip32_index = LedgerBIP32Index {
+            bip32_change: bip32_chain[3] | Segment::HARDEN_MASK,
+            bip32_index: bip32_chain[4] | Segment::HARDEN_MASK,
+        };
+
+        // Lock the mutex to prevent multiple simultaneous requests to a ledger.
+        let lock = self.mutex.lock().await;
+
+        let ledger = get_ledger(coin_type, account_index, self.is_simulator).map_err(Error::from)?;
+
+        log::debug!("[LEDGER] prepare_blind_signing");
+        log::debug!("[LEDGER] {:?} {:?}", bip32_index, msg);
+        ledger
+            .prepare_blind_signing(vec![bip32_index], msg)
+            .map_err(Error::from)?;
+
+        // Show essence to user, if denied by user, it returns with `DeniedByUser` Error.
+        log::debug!("[LEDGER] await user confirmation");
+        ledger.user_confirm().map_err(Error::from)?;
+
+        // Sign.
+        let signature_bytes = ledger.sign(1).map_err(Error::from)?;
+
+        drop(ledger);
+        drop(lock);
+
+        let mut unpacker = SliceUnpacker::new(&signature_bytes);
+
+        // Unpack and return signature.
+        return match Unlock::unpack::<_, true>(&mut unpacker, &())? {
+            Unlock::Signature(SignatureUnlock(Signature::Ed25519(signature))) => Ok(signature),
+            _ => Err(Error::UnsupportedOperation.into()),
+        };
     }
 
     async fn sign_secp256k1_ecdsa(
@@ -189,21 +237,21 @@ impl SecretManage for LedgerSecretManager {
         prepared_transaction: &PreparedTransactionData,
         time: Option<u32>,
     ) -> Result<Unlocks, <Self as SecretManage>::Error> {
-        let mut input_bip32_indices: Vec<LedgerBIP32Index> = Vec::new();
-        let mut coin_type: Option<u32> = None;
-        let mut account_index: Option<u32> = None;
+        let mut input_bip32_indices = Vec::new();
+        let mut coin_type = None;
+        let mut account_index = None;
 
         let input_len = prepared_transaction.inputs_data.len();
 
         for input in &prepared_transaction.inputs_data {
-            let bip32_indices: Vec<u32> = match &input.chain {
+            let bip32_indices = match &input.chain {
                 Some(chain) => {
                     chain
                         .segments()
                         .iter()
                         // XXX: "ser32(i)". RTFSC: [crypto::keys::slip10::Segment::from_u32()]
                         .map(|seg| u32::from_be_bytes(seg.bs()))
-                        .collect()
+                        .collect::<Vec<_>>()
                 }
                 None => return Err(Error::MissingBip32Chain)?,
             };
@@ -254,14 +302,14 @@ impl SecretManage for LedgerSecretManager {
             let (remainder_address, remainder_bip32): (Option<&Address>, LedgerBIP32Index) =
                 match &prepared_transaction.remainder {
                     Some(a) => {
-                        let remainder_bip32_indices: Vec<u32> = match &a.chain {
+                        let remainder_bip32_indices = match &a.chain {
                             Some(chain) => {
                                 chain
                                     .segments()
                                     .iter()
                                     // XXX: "ser32(i)". RTFSC: [crypto::keys::slip10::Segment::from_u32()]
                                     .map(|seg| u32::from_be_bytes(seg.bs()))
-                                    .collect()
+                                    .collect::<Vec<_>>()
                             }
                             None => return Err(Error::MissingBip32Chain.into()),
                         };